# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: HHC LINE Function Bot API
  version: 1.0.0
  description: >-
    HTTP contract for LINE webhooks, service probes, and administrator media-sync mechanics.
    LINE chat commands, controlled functions, postbacks, and administrator chat commands are not
    separate HTTP APIs.
servers:
  - url: http://hhc-line-function-bot
    description: Internal Dapr service address; gateway-facing traffic is forwarded to this service.
...
paths:
...
  /api/line/media-sync/collections:
    get:
      operationId: listMediaSyncCollections
      summary: List managed LINE media collections
      description: >-
        admin-fe calls through API Gateway. The service accepts only the configured api-gateway Dapr
        caller and API token, validates the gateway-trusted HHC user UUID, and asks Account API to
        verify that user has media-sync administration permission before calling Asset API.
      tags: [Admin]
      x-hhc-visibility: admin
      x-hhc-callers: [admin-fe]
      security:
        - gatewayCaller: []
          appApiToken: []
          trustedUser: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Managed collection page.
          headers: { X-HHC-Request-ID: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedCollectionPage" }
        <<: *assetErrorResponses
    post:
      operationId: createMediaSyncCollection
      summary: Create a managed LINE media collection
      description: Creates an Asset-owned collection after the shared media-sync authorization gate.
      tags: [Admin]
      x-hhc-visibility: admin
      x-hhc-callers: [admin-fe]
      security:
        - gatewayCaller: []
          appApiToken: []
          trustedUser: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CollectionNameRequest" }
      responses:
        "201":
          description: Collection created.
          headers: { X-HHC-Request-ID: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
        <<: *assetErrorResponses
  /api/line/media-sync/acl-subjects:
    get:
      operationId: searchMediaSyncAclSubjects
      summary: Search eligible media-sync ACL subjects
      description: >-
        Runs behind the shared media-sync authorization gate, then asks Account API for a bounded
        page of users or roles. The trusted user identity is supplied separately as the requesting
        administrator and cannot be selected by the request body.
      tags: [Admin]
      x-hhc-visibility: admin
      x-hhc-callers: [admin-fe]
      security:
        - gatewayCaller: []
          appApiToken: []
          trustedUser: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: subjectType
          in: query
          required: true
          schema: { type: string, enum: [user, role] }
        - name: q
          in: query
          required: true
          description: Trimmed before use; the normalized value is at most 120 UTF-8 bytes and has no control characters.
          schema:
            type: string
            maxLength: 120
            not: { pattern: '\p{Cc}|[\uD800-\uDFFF]' }
        - name: page
          in: query
          required: true
          schema: { type: integer, minimum: 1, maximum: 1000 }
        - name: perPage
          in: query
          required: true
          schema: { type: integer, minimum: 1, maximum: 50 }
      responses:
        "200":
          description: Eligible ACL subject page.
          headers: { X-HHC-Request-ID: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AclSubjectPage" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/line/media-sync/collections/{collectionId}:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
    patch:
      operationId: renameMediaSyncCollection
      summary: Rename a managed LINE media collection
      description: Renames the Asset-owned collection after the shared media-sync authorization gate.
      tags: [Admin]
      x-hhc-visibility: admin
      x-hhc-callers: [admin-fe]
      security:
        - gatewayCaller: []
          appApiToken: []
          trustedUser: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CollectionNameRequest" }
      responses:
        "200":
          description: Collection renamed.
          headers: { X-HHC-Request-ID: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
        <<: *assetErrorResponses
    delete:
      operationId: deleteMediaSyncCollection
      summary: Delete a managed LINE media collection
      description: Starts the binding-safe collection deletion lifecycle before deleting it in Asset API.
      tags: [Admin]
      x-hhc-visibility: admin
      x-hhc-callers: [admin-fe]
      security:
        - gatewayCaller: []
          appApiToken: []
          trustedUser: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Collection deleted.
          headers: { X-HHC-Request-ID: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
        <<: *assetErrorResponses
  /api/line/media-sync/collections/{collectionId}/items:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
    get:
      operationId: listMediaSyncCollectionItems
      summary: List managed collection items
      description: Lists Asset-owned collection items after the shared media-sync authorization gate.
      tags: [Admin]
      x-hhc-visibility: admin
      x-hhc-callers: [admin-fe]
      security:
        - gatewayCaller: []
          appApiToken: []
          trustedUser: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: q
          in: query
          description: Optional search text, at most 255 UTF-8 bytes with no control characters.
          schema:
            type: string
            maxLength: 255
            not: { pattern: '\p{Cc}|[\uD800-\uDFFF]' }
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Managed collection item page.
          headers: { X-HHC-Request-ID: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedCollectionItemPage" }
        <<: *assetErrorResponses
  /api/line/media-sync/collections/{collectionId}/retention:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
    patch:
      operationId: updateMediaSyncCollectionRetention
      summary: Update collection retention
      description: Sets Asset-owned collection retention after the shared media-sync authorization gate.
      tags: [Admin]
      x-hhc-visibility: admin
      x-hhc-callers: [admin-fe]
      security:
        - gatewayCaller: []
          appApiToken: []
          trustedUser: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RetentionRequest" }
      responses:
        "200":
          description: Collection retention updated.
          headers: { X-HHC-Request-ID: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
        <<: *assetErrorResponses
  /api/line/media-sync/collections/{collectionId}/items/{itemId}:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
      - $ref: "#/components/parameters/ItemId"
    patch:
      operationId: renameMediaSyncCollectionItem
      summary: Rename a collection item
      description: Renames an Asset-owned collection item after the shared authorization gate.
      tags: [Admin]
      x-hhc-visibility: admin
      x-hhc-callers: [admin-fe]
      security:
        - gatewayCaller: []
          appApiToken: []
          trustedUser: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ItemNameRequest" }
      responses:
        "200":
          description: Collection item renamed.
          headers: { X-HHC-Request-ID: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManagedCollectionItem" }
        <<: *assetErrorResponses
  /api/line/media-sync/collections/{collectionId}/items/retention:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
    post:
      operationId: setMediaSyncItemsRetention
      summary: Set item retention exemptions
      description: Updates retention exemption for a bounded item set after the shared authorization gate.
      tags: [Admin]
      x-hhc-visibility: admin
      x-hhc-callers: [admin-fe]
      security:
        - gatewayCaller: []
          appApiToken: []
          trustedUser: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ItemsRetentionRequest" }
      responses:
        "200":
          description: Item retention exemptions updated.
          headers: { X-HHC-Request-ID: { $ref: "#/components/headers/RequestId" } }
        <<: *assetErrorResponses
  /api/line/media-sync/collections/{collectionId}/items/delete:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
    post:
      operationId: deleteMediaSyncCollectionItems
      summary: Delete collection items
      description: Deletes a bounded Asset-owned item set after the shared authorization gate.
      tags: [Admin]
      x-hhc-visibility: admin
      x-hhc-callers: [admin-fe]
      security:
        - gatewayCaller: []
          appApiToken: []
          trustedUser: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ItemIdsRequest" }
      responses:
        "200":
          description: Item deletion counts.
          headers: { X-HHC-Request-ID: { $ref: "#/components/headers/RequestId" } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteItemsResponse" }
        <<: *assetErrorResponses
  /api/line/media-sync/collections/{collectionId}/items/content-tickets:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
    post:
      operationId: issueMediaSyncContentTickets
      summary: Issue short-lived item content tickets
      description: >-
        Issues private, no-store Asset content tickets for a bounded item set after the shared
        authorization gate. Ticket values are response-only credentials and are not examples or
        durable API configuration.
      tags: [Admin]
      x-hhc-visibility: admin
      x-hhc-callers: [admin-fe]
      security:
        - gatewayCaller: []
          appApiToken: []
          trustedUser: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ItemIdsRequest" }
      responses:
        "201":
          description: Short-lived content ticket batch.
          headers:
            X-HHC-Request-ID: { $ref: "#/components/headers/RequestId" }
            Cache-Control: { schema: { type: string, const: "private, no-store" } }
            Referrer-Policy: { schema: { type: string, const: no-referrer } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContentTicketBatch" }
        <<: *assetErrorResponses
...
components:
  securitySchemes:
    lineSignature:
      type: apiKey
      in: header
      name: X-Line-Signature
      description: LINE HMAC-SHA256 signature verified against the unmodified request bytes.
    gatewayCaller:
      type: apiKey
      in: header
      name: Dapr-Caller-App-Id
      description: Must exactly match the configured api-gateway Dapr app id.
    appApiToken:
      type: apiKey
      in: header
      name: dapr-api-token
      description: Shared app API token compared without exposing its value.
    trustedUser:
      type: apiKey
      in: header
      name: X-HHC-User-ID
      description: >-
        Gateway-trusted HHC user UUID. After caller and token checks, every media-sync request is
        authorized by Account API for the media-sync administration permission and fails closed.
  parameters:
...
    RequestId:
      name: X-HHC-Request-ID
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^\S(?:.*\S)?$'
        not: { pattern: '\p{Cc}|[\uD800-\uDFFF]' }
      description: >-
        Already-trimmed value of at most 128 UTF-8 bytes with no control characters. A valid value
        is reused for downstream correlation; otherwise the service generates one.
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string, minLength: 1, maxLength: 2048 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
...